--- ExplorersIcebox/Scheduler/Tasks/Task_Interact.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ExplorersIcebox.Scheduler.Tasks
-{
-    internal static class Task_Interact
-    {
-        public static void Enqueue()
-        {
-            // foreach (Svc.Objects)
-        }
-    }
-}
