+++ ExplorersIcebox/Util/ItemData.cs
@@ -53,53 +53,60 @@ public static class ItemData
     public const int HawksEyeSand_ID = 41633;
     public const int CrystalFormation_ID = 41634;
 
-    public static Dictionary<int, string> IslandItems = new()
+    public class IslandItemInfo
     {
-        { 37551, "Palm Leaf" },
-        { 37552, "Apple" },
-        { 37553, "Branch" },
-        { 37554, "Stone" },
-        { 37555, "Clam" },
-        { 37556, "Laver" },
-        { 37557, "Coral" },
-        { 37558, "Islewort" },
-        { 37559, "Sand" },
-        { 37560, "Log" },
-        { 37561, "Palm Log" },
-        { 37562, "Vine" },
-        { 37563, "Sap" },
-        { 37564, "Copper" },
-        { 37565, "Limestone" },
-        { 37566, "Rock Salt" },
-        { 37567, "Sugarcane" },
-        { 37568, "Cotton" },
-        { 37569, "Hemp" },
-        { 37570, "Clay" },
-        { 37571, "Tinsand" },
-        { 37572, "Iron Ore" },
-        { 37573, "Quartz" },
-        { 37574, "Leucogranite" },
-        { 37575, "Islefish" },
-        { 37576, "Squid" },
-        { 37577, "Jellyfish" },
-        { 39224, "Resin" },
-        { 39225, "Coconut" },
-        { 39226, "Beehive" },
-        { 39227, "Wood Opal" },
-        { 39228, "Multicolored Isleblooms" },
-        { 39887, "Coal" },
-        { 39888, "Shale" },
-        { 39889, "Glimshroom" },
-        { 39890, "Marble" },
-        { 39891, "Mythril Ore" },
-        { 39892, "Effervescent Water" },
-        { 39893, "Spectrine" },
-        { 41630, "Durium Sand" },
-        { 41631, "Yellow Copper Ore" },
-        { 41632, "Gold Ore" },
-        { 41633, "Hawk's Eye Sand" },
-        { 41634, "Crystal Formation" },
+        public string ItemName { get; set; } = "";
+        public int SellSlot { get; set; } = 0;
+    }
+
+    public static Dictionary<int, IslandItemInfo> IslandItems = new()
+    {
+        { 37551, new IslandItemInfo { ItemName = "Palm Leaf", SellSlot = 0 } },
+        { 37552, new IslandItemInfo { ItemName = "Apple", SellSlot = 0 } },
+        { 37553, new IslandItemInfo { ItemName = "Branch", SellSlot = 0 } },
+        { 37554, new IslandItemInfo { ItemName = "Stone", SellSlot = 0 } },
+        { 37555, new IslandItemInfo { ItemName = "Clam", SellSlot = 0 } },
+        { 37556, new IslandItemInfo { ItemName = "Laver", SellSlot = 0 } },
+        { 37557, new IslandItemInfo { ItemName = "Coral", SellSlot = 0 } },
+        { 37558, new IslandItemInfo { ItemName = "Islewort", SellSlot = 0 } },
+        { 37559, new IslandItemInfo { ItemName = "Sand", SellSlot = 0 } },
+        { 37560, new IslandItemInfo { ItemName = "Log", SellSlot = 0 } },
+        { 37561, new IslandItemInfo { ItemName = "Palm Log", SellSlot = 0 } },
+        { 37562, new IslandItemInfo { ItemName = "Vine", SellSlot = 0 } },
+        { 37563, new IslandItemInfo { ItemName = "Sap", SellSlot = 0 } },
+        { 37564, new IslandItemInfo { ItemName = "Copper", SellSlot = 0 } },
+        { 37565, new IslandItemInfo { ItemName = "Limestone", SellSlot = 0 } },
+        { 37566, new IslandItemInfo { ItemName = "Rock Salt", SellSlot = 0 } },
+        { 37567, new IslandItemInfo { ItemName = "Sugarcane", SellSlot = 0 } },
+        { 37568, new IslandItemInfo { ItemName = "Cotton", SellSlot = 0 } },
+        { 37569, new IslandItemInfo { ItemName = "Hemp", SellSlot = 0 } },
+        { 37570, new IslandItemInfo { ItemName = "Clay", SellSlot = 0 } },
+        { 37571, new IslandItemInfo { ItemName = "Tinsand", SellSlot = 0 } },
+        { 37572, new IslandItemInfo { ItemName = "Iron Ore", SellSlot = 0 } },
+        { 37573, new IslandItemInfo { ItemName = "Quartz", SellSlot = 0 } },
+        { 37574, new IslandItemInfo { ItemName = "Leucogranite", SellSlot = 0 } },
+        { 37575, new IslandItemInfo { ItemName = "Islefish", SellSlot = 0 } },
+        { 37576, new IslandItemInfo { ItemName = "Squid", SellSlot = 0 } },
+        { 37577, new IslandItemInfo { ItemName = "Jellyfish", SellSlot = 0 } },
+        { 39224, new IslandItemInfo { ItemName = "Resin", SellSlot = 0 } },
+        { 39225, new IslandItemInfo { ItemName = "Coconut", SellSlot = 0 } },
+        { 39226, new IslandItemInfo { ItemName = "Beehive", SellSlot = 0 } },
+        { 39227, new IslandItemInfo { ItemName = "Wood Opal", SellSlot = 0 } },
+        { 39228, new IslandItemInfo { ItemName = "Multicolored Isleblooms", SellSlot = 0 } },
+        { 39887, new IslandItemInfo { ItemName = "Coal", SellSlot = 0 } },
+        { 39888, new IslandItemInfo { ItemName = "Shale", SellSlot = 0 } },
+        { 39889, new IslandItemInfo { ItemName = "Glimshroom", SellSlot = 0 } },
+        { 39890, new IslandItemInfo { ItemName = "Marble", SellSlot = 0 } },
+        { 39891, new IslandItemInfo { ItemName = "Mythril Ore", SellSlot = 0 } },
+        { 39892, new IslandItemInfo { ItemName = "Effervescent Water", SellSlot = 0 } },
+        { 39893, new IslandItemInfo { ItemName = "Spectrine", SellSlot = 0 } },
+        { 41630, new IslandItemInfo { ItemName = "Durium Sand", SellSlot = 0 } },
+        { 41631, new IslandItemInfo { ItemName = "Yellow Copper Ore", SellSlot = 0 } },
+        { 41632, new IslandItemInfo { ItemName = "Gold Ore", SellSlot = 0 } },
+        { 41633, new IslandItemInfo { ItemName = "Hawk's Eye Sand", SellSlot = 0 } },
+        { 41634, new IslandItemInfo { ItemName = "Crystal Formation", SellSlot = 0 } },
     };
+
 
     public class GatheringNode
     {
